:memo: Parser - Add documentation
