+++ packages/ntm_parser/README.md
@@ -1,1 +1,3 @@
 ## ntm_parser
+
+A ntm parser that produces the corresponding syntax tree from a valid sequence of tokens.
