+++ packages/ntm_parser/lib/ntm_parser.dart
@@ -1,3 +1,5 @@
 library ntm_parser;
 
+export 'src/parse_error.dart';
+export 'src/parse_result.dart';
 export 'src/parser.dart';
