--- packages/ntm_parser/lib/src/ntm_parser_base.dart
@@ -1,6 +1,0 @@
-// TODO: Put public facing types in this file.
-
-/// Checks if you are awesome. Spoiler: you are.
-class Awesome {
-  bool get isAwesome => true;
-}
